+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules
 node_modules
 .env
+.next
 
 api-search.py
