--- src/index.js
@@ -1,18 +1,0 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import './styles/index.css'
-import 'react-toastify/dist/ReactToastify.css'
-import App from './App'
-
-/* react-query */
-/* import { QueryClient, QueryClientProvider } from 'react-query'
-import { ReactQueryDevtools } from 'react-query/devtools'
-
-const queryClient = new QueryClient() */
-
-const root = ReactDOM.createRoot(document.getElementById('root'))
-root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-)
